pyproject.toml
[project]
name = "agentic-llm-for-better-results"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "alpaca-py>=0.31",
  "alpha-vantage>=3",
  "cohere>=5.9.2",
  "crewai>=0.65.2",
  "crewai-tools>=0.12.1",
  "datasets>=3",
  "diskcache>=5.6.3",
  "edgartools>=2.34",
  "faiss-cpu>=1.9",
  "feedparser>=6.0.11",
  "google-generativeai>=0.8.2",
  "great-tables>=0.11.1",
  "html-to-markdown>=1.1",
  "html2text>=2024.2.26",
  "huggingface-hub[cli]",
  "hvplot>=0.11",
  "ipywidgets>=8.1.5",
  "jupyterlab>=4.2.5",
  "jupyterlab-quarto>=0.3.5",
  "langchain>=0.2.16",
  "langchain-huggingface>=0.0.3",
  "litellm>=1.48.6",
  "llama-index>=0.11.16",
  "llama-index-embeddings-huggingface>=0.3.1",
  "llama-index-llms-groq>=0.2",
  "llama-index-llms-huggingface-api>=0.2",
  "llama-index-llms-ollama>=0.3.3",
  "llama-index-readers-json>=0.2",
  "matplotlib>=3.9.2",
  "nbconvert>=7.16.4",
  "pandas>=2.2.2",
  "panel>=1.5.1",
  "plotly>=5.24.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.5.2",
  "seaborn>=0.13.2",
  "sentence-transformers>=3.1.1",
  "tabulate>=0.9",
  "torch>=2.4.1",
  "tqdm>=4.66.5",
  "transformers>=4.44.2",
  "typer>=0.12.5",
  "yfinance>=0.2.43",
]

optional-dependencies.pdf = [
  "pdfkit>=1",
  "pypdf>=4.3.1",
]
optional-dependencies.sec = [
  "datamule>=0.332",
  "edgartools>=2.34",
]

[tool.ruff]
line-length = 100
lint.extend-per-file-ignores."__init__.py" = [
  "F401",
  "F403",
]

lint.unfixable = [
  "F401", # F401 unused imports
]

[tool.uv]
dev-dependencies = [
  "mypy>=1.12.0",
  "pandas-stubs>=2.2.3.241009",
  "pre-commit>=3.8.0",
]
